Implemented uniform analog pin sensor interface

+++ src/sensors/Sensor.cpp
@@ -4,12 +4,12 @@
 
 using namespace wlp;
 
-void Sensor::value_read(float value) {
+void Sensor::value_read(double value) {
     UNUSED(value);
 }
 
-float Sensor::read() {
-    float value = read_value();
+double Sensor::read() {
+    double value = read_value();
     value_read(value);
     return value;
 }
