+++ src/sensors/Sensor.h
@@ -2,17 +2,13 @@
 #define GOOSE_SENSORS_SENSOR_H
 
 namespace wlp {
-
     class Sensor {
     protected:
-        virtual float read_value() = 0;
-
-        virtual void value_read(float value);
-
+        virtual double read_value() = 0;
+        virtual void value_read(double value);
     public:
-        float read();
+        double read();
     };
-
 }
 
 #endif //GOOSE_SENSORS_SENSOR_H
